+++ suomen-vaylat-app/src/components/app-info-modal/AppInfoModalContent.jsx
@@ -2,8 +2,9 @@ import React, { useRef, useState, useEffect} from 'react';
 import styled from 'styled-components';
 import strings from '../../translations';
 import { getAppBuildDate, getAppVersion } from '../../utils/appInfoUtil';
-import { isMobile} from '../../theme/theme';
+import { isMobile, size} from '../../theme/theme';
 import { motion, AnimatePresence } from 'framer-motion';
+import { useSelector } from 'react-redux';
 
 import { Swiper, SwiperSlide } from 'swiper/react';
 
@@ -201,11 +202,19 @@ export const AppInfo = () => {
 }
 
 export const VersionInfo = ({currentAppVersion, currentAppBuildDate}) => {
+    const { channel } = useSelector(state => state.rpc);
+    const [oskariVersion, setOskariVersion] = useState(); 
+    
+    channel?.getInfo(function(oskariInfo) {
+        setOskariVersion(oskariInfo.version); 
+    });
+
     return (
         <div>
             {isMobile && <StyledHeading>{strings.appInfo.versionInfo.title}</StyledHeading>}
             <StyledTitle><p>{strings.appInfo.versionInfo.appVersion + currentAppVersion}</p></StyledTitle>
             <StyledTitle><p>{strings.appInfo.versionInfo.appLastUpdate + currentAppBuildDate}</p></StyledTitle>
+            <StyledTitle><p>{strings.appInfo.versionInfo.oskari} {oskariVersion}</p></StyledTitle>
         </div>
     );
 };
@@ -245,7 +254,9 @@ export const SourcesAndTermsOfUse = () => {
             <div>
                 <StyledHeading> {strings.appInfo.dataSources.title}</StyledHeading>
                 <p>{strings.appInfo.dataSources.municipalityImage} <StyledLink href={strings.appInfo.dataSources.municipalityLink}>{strings.appInfo.dataSources.municipalityLink}</StyledLink> </p>
-                <p>{strings.appInfo.dataSources.landSurvey}</p>
+                <p>{strings.appInfo.dataSources.landSurvey} <StyledLink href={strings.appInfo.dataSources.landSurveyLink}>{strings.appInfo.dataSources.landSurveyLink}</StyledLink> </p>
+                <p>{strings.appInfo.dataSources.roadmaps}</p>
+                <p>{strings.appInfo.dataSources.syke}</p>
             </div>
             <div>
                 <StyledHeading>{strings.appInfo.termsOfUse.title}</StyledHeading>
@@ -263,6 +274,8 @@ export const AppInfoModalContent = () => {
     const inputEl = useRef(null);
     const [isNavOpen, setIsNavOpen] = useState(true);
 
+    const showMobileView = isMobile && window.innerWidth < parseInt(size.mobileL);
+
     // App build info
     const currentAppVersion = getAppVersion();
     const currentAppBuildDate = getAppBuildDate();
@@ -270,7 +283,7 @@ export const AppInfoModalContent = () => {
     const [tabIndex, setTabIndex] = useState(0);
 
     useEffect(() => {
-        isMobile && setTabIndex(-1);
+        showMobileView && setTabIndex(-1);
     }, [])
 
     const tabsContent = [
@@ -325,11 +338,11 @@ export const AppInfoModalContent = () => {
     return (
         <>
             <StyledContent>
-                {isMobile && 
+                {showMobileView && 
                 <>
                     <StyledMenuContainer key="menuContainer">
                         <StyledCloseMenuContainer>
-                            <StyledCloseMenuIcon onClick={() => setIsNavOpen(!isNavOpen)} icon={!isNavOpen && faArrowLeft} />
+                            {!isNavOpen && <StyledCloseMenuIcon onClick={() => setIsNavOpen(!isNavOpen)} icon={faArrowLeft} />}
                         </StyledCloseMenuContainer>
                     </StyledMenuContainer>
                     <StyledMobileContainer
@@ -371,7 +384,7 @@ export const AppInfoModalContent = () => {
                 </>
                 }
 
-                {!isMobile &&
+                {!showMobileView &&
                     <StyledTabs
                     tabIndex={tabIndex}
                     tabsCount={tabsContent.length}
@@ -403,7 +416,7 @@ export const AppInfoModalContent = () => {
                         setTabIndex(e.activeIndex);
                     }}
                     allowTouchMove={false} // Disable swiping
-                    speed={isMobile? 0 : 300}
+                    speed={showMobileView? 0 : 300}
                 >
                     {
                         tabsContent.map((tab, index) => {
@@ -413,10 +426,10 @@ export const AppInfoModalContent = () => {
                                     key={'ai_tab_content_' + index}
                                 >
                                     <AnimatePresence>
-                                        {isMobile && !isNavOpen ? 
-                                        <StyledMobileTabContent key={'tabContent_' + index} variants={isMobile && variants} initial="initial" animate={isNavOpen ? "hidden" : "visible"} exit="exit">
+                                        {showMobileView && !isNavOpen ? 
+                                        <StyledMobileTabContent key={'tabContent_' + index} variants={showMobileView && variants} initial="initial" animate={isNavOpen ? "hidden" : "visible"} exit="exit">
                                         {tab.content}
-                                        </StyledMobileTabContent> : !isMobile && tab.content
+                                        </StyledMobileTabContent> : !showMobileView && tab.content
                                     }
                                     </AnimatePresence>
                                     </SwiperSlide>