+++ ReleaseNotes.md
@@ -1,8 +1,20 @@
 # Release Notes
 
+## 1.6.0
+
+For a full list of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/9?closed=1
+
+* Fix backgroundmap opacities not updating
+* Tablet style fixes and improvements & crash handling
+* Reset geometries on logo click
+* Fix removed geometries staying on the map
+* Localization updates
+* Download improvements
+
 ## 1.5.0
 
 For a full list of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/8?closed=1
+
 * Backgroundmap grouping recursive
 * Open legend when opening a themelink
 * Fix GFI error when layer some property contains JSONObject values
@@ -10,6 +22,7 @@ For a full list of changes see: https://github.com/finnishtransportagency/suomen
 ## 1.4.0
 
 For a full list of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/7?closed=1
+
 * Bug fixes
 * Markers